Fix rz-ghidra build due to `ht_*` move
