+++ librz/include/rz_lib.h
@@ -3,7 +3,7 @@
 
 #include "rz_types.h"
 #include "rz_list.h"
-#include <ht_pu.h>
+#include <rz_util/ht_pu.h>
 
 #if __UNIX__
 #include <dlfcn.h>
