+++ librz/include/rz_list.h
@@ -2,7 +2,7 @@
 #define RZ_LIST_H
 
 #include <rz_types.h>
-#include <ht_pp.h>
+#include <rz_util/ht_pp.h>
 
 #ifdef __cplusplus
 extern "C" {
