+++ librz/include/rz_util/rz_event.h
@@ -8,7 +8,7 @@
 extern "C" {
 #endif
 
-#include <ht_up.h>
+#include "ht_up.h"
 #include <rz_vector.h>
 
 typedef struct rz_event_t {
