+++ librz/include/rz_util/rz_th_ht.h
@@ -6,10 +6,10 @@
 #define RZ_TH_HASH_TABLE_H
 
 #include <rz_th.h>
-#include <ht_pp.h>
-#include <ht_up.h>
-#include <ht_uu.h>
-#include <ht_pu.h>
+#include "ht_pp.h"
+#include "ht_up.h"
+#include "ht_uu.h"
+#include "ht_pu.h"
 
 #ifdef __cplusplus
 extern "C" {
